Added SD logging & LED
Needed to remove the GPS library to free up RAM on the UNO.
We switched to using a softserial line and reading in one line
at a time. The line is the checked for validity, then if there's
a fix, the data is logged along with the PM2.5 data.

Also added LED capabilities for live readout of AQI range
as well as the current status of the sensors.